+++ inst/extdata/IGVSnapshot.R
@@ -31,7 +31,7 @@
 #' @details This function depend on SRAdb startIGV IGVsocket IGVgenome IGVload IGVgoto IGVsnapshot IGVsession
 #' @examples
 #' if(interactive()){
-#' #library(SRAdb)
+#' library(SRAdb)
 #' exampleBams = file.path(system.file('extdata',package='SRAdb'),
 #'              dir(system.file('extdata',package='SRAdb'),pattern='bam$'))
 #' IGVSnapshot(maxMem="mm", genomeBuild="hg18", bamFileFullPathOrURLs= exampleBams,